django/requirements.txt
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.4
django-classy-tags==1.0.0
django-taggit==1.2.0
django-taggit-templatetags2==1.6.1
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
gunicorn==19.9.0
httpie==2.0.0
idna==2.8
isort==4.3.21
mccabe==0.6.1
mysqlclient==1.4.4
Pillow==7.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
uWSGI==2.0.18
whitenoise==4.1.3
